requirements.txt
coverage==6.4.1
celery==5.2.7
Django==4.0.5
django-cors-headers==3.13.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djangorestframework-recursive==0.1.2
drf-yasg==1.20.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.4.0
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
channels==3.0.4
channels_redis==3.4.0
asgiref==3.5.2
Pillow==9.1.1
redis==4.3.4